requirements.txt
streamlit
openai
python-dotenv
supabase
streamlit>=1.32
openai>=1.0
supabase>=1.0
python-dotenv>=1.0